Move docker daemon config to a separate struct under dockerdaemon package (#233)
