+++ agent/cmd/cmd.go
@@ -128,7 +128,6 @@ func Run(flags *Flags, opts ...Option) {
 			}
 		}
 	}
-	config = config.applyDefaults()
 
 	if overrides.logger != nil {
 		log.SetGlobalLogger(overrides.logger.Sugar())
@@ -205,8 +204,7 @@ func Run(flags *Flags, opts ...Option) {
 	}
 
 	registryAddr := fmt.Sprintf("127.0.0.1:%d", flags.AgentRegistryPort)
-	dockerCli, err := dockerdaemon.NewDockerClient(
-		config.DockerHost, config.DockerScheme, config.DockerClientVersion, registryAddr)
+	dockerCli, err := dockerdaemon.NewDockerClient(config.DockerDaemon, registryAddr)
 	if err != nil {
 		log.Fatalf("failed to init docker client for preload: %s", err)
 	}
