+++ agent/cmd/config.go
@@ -16,6 +16,7 @@ package cmd
 import (
 	"github.com/uber/kraken/agent/agentserver"
 	"github.com/uber/kraken/core"
+	"github.com/uber/kraken/lib/dockerdaemon"
 	"github.com/uber/kraken/lib/dockerregistry"
 	"github.com/uber/kraken/lib/store"
 	"github.com/uber/kraken/lib/torrent/networkevent"
@@ -44,22 +45,5 @@ type Config struct {
 	Nginx           nginx.Config                   `yaml:"nginx"`
 	TLS             httputil.TLSConfig             `yaml:"tls"`
 	AllowedCidrs    []string                       `yaml:"allowed_cidrs"`
-
-	// Configs For connecting to docker daemon for preload.
-	DockerHost          string `yaml:"docker_host"`
-	DockerScheme        string `yaml:"docker_scheme"`
-	DockerClientVersion string `yaml:"docker_version"`
-}
-
-func (c Config) applyDefaults() Config {
-	if c.DockerHost == "" {
-		c.DockerHost = "unix:///var/run/docker.sock"
-	}
-	if c.DockerScheme == "" {
-		c.DockerScheme = "http"
-	}
-	if c.DockerClientVersion == "" {
-		c.DockerClientVersion = "1.21"
-	}
-	return c
+	DockerDaemon    dockerdaemon.Config            `yaml:docker_daemon`
 }
