+++ lib/dockerdaemon/config.go
@@ -1,0 +1,34 @@
+// Copyright (c) 2016-2019 Uber Technologies, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package dockerdaemon
+
+// Configs connecting to docker daemon.
+type Config struct {
+	DockerHost          string `yaml:"docker_host"`
+	DockerScheme        string `yaml:"docker_scheme"`
+	DockerClientVersion string `yaml:"docker_version"`
+}
+
+func (c Config) applyDefaults() Config {
+	if c.DockerHost == "" {
+		c.DockerHost = "unix:///var/run/docker.sock"
+	}
+	if c.DockerScheme == "" {
+		c.DockerScheme = "http"
+	}
+	if c.DockerClientVersion == "" {
+		c.DockerClientVersion = "1.21"
+	}
+	return c
+}
